pyproject.toml
[project]
name = "mcp-example"
version = "0.1.0"
description = "Tiny example of MCP use"
authors = [
    {name = "Andrew Savchenko",email = "andrew@savchenko.net"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "pydantic-ai-slim[fastmcp,openai] (>=1.72.0,<2.0.0)",
    "mcp[cli] (>=1.26.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
